hyperliquid/audit fixes (#1700)

+++ examples/oft-hyperliquid/contracts/MyHyperLiquidComposer.sol
@@ -13,6 +13,6 @@ contract MyHyperLiquidComposer is HyperLiquidComposer {
     constructor(
         address _oft,
         uint64 _hlIndexId,
-        int64 _assetDecimalDiff
+        int8 _assetDecimalDiff
     ) HyperLiquidComposer(_oft, _hlIndexId, _assetDecimalDiff) {}
 }
