+++ packages/hyperliquid-composer/test/mocks/FeeTokenMock.sol
@@ -8,7 +8,7 @@ contract FeeTokenMock is FeeToken {
     constructor(
         address _oft,
         uint64 _coreIndexId,
-        int64 _assetDecimalDiff
+        int8 _assetDecimalDiff
     ) HyperLiquidComposer(_oft, _coreIndexId, _assetDecimalDiff) {}
 
     function createRawActionPayloadERC20(address _to, uint64 _coreAmount) public view returns (bytes memory payload) {
